+++ NEWS.md
@@ -2,16 +2,17 @@
 
 Version numbers follow [Semantic Versioning](https://semver.org/).
 
-# portalcasting 0.10.0
-*Active development*
+# [portalcasting 0.10.0](https://github.com/weecology/portalcasting/releases/tag/v0.10.0)
+*2019-09-13*
 
 ### Model evaluation and ensembling added back in
 * Were removed with the updated version from 0.8.1 to 0.9.0 to allow time to develop the code with the new infrastructure. 
 * Model evaluation happens within the cast tab output as before.
 
 ### Temporarily removed figures returned
-* Associated with the evaluation and ensembling.
+* Associated with the evaluation.
 * Plotting of error as a function of lead time for multiple species and multiple models. Now has a fall-back arrangement that works for a single species-model combination.
+* Plotting RMSE and coverage within species-model combinations.
 
 ### Flexing model controls to allow user-defined lists for prefab models
 * For sandboxing with existing models, it is useful to be able to change a parameter in the model's controls, such as the data sets. Previously, that would require a lot of hacking around. Now, it's as simple as inputting the desired controls and flipping `arg_checks = FALSE`.